# swagger.yaml
openapi: 3.0.0
info:
  title: Blog API
  description: API documentation for Blog
  version: 1.0.0
servers:
  - url: /api
    description: API base URL
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ImageMedia:
      type: object
      properties:
        title:
          type: string
        altText:
          type: string
        url:
          type: string
          format: url
        type:
          type: string
          enum: [image, video, document]
        size:
          type: string
        dimensions:
          type: string
      example:
        title: "Node.js Architecture"
        altText: "Diagram showing Node.js event loop"
        url: "https://example.com/images/nodejs-arch.jpg"
        type: "image"
        size: "245.5 KB"
        dimensions: "1920x1080"
paths:
...
  /images/upload:
    post:
      tags:
        - Image
      summary: Upload new image
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                title:
                  type: string
                altText:
                  type: string
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  url:
                    type: string
                  data:
                    $ref: '#/components/schemas/ImageMedia'
                  message:
                    type: string
                example:
                  success: true
                  url: "https://example.com/images/uploaded-image.jpg"
                  data:
                    title: "Uploaded Image"
                    altText: "Description of uploaded image"
                    url: "https://example.com/images/uploaded-image.jpg"
                    type: "image"
                    size: "156.7 KB"
                    dimensions: "1024x768"
                  message: "Image uploaded successfully"
  /images/{id}:
    put:
      tags:
        - Image
      summary: Update image details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: objectId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - altText
              properties:
                title:
                  type: string
                altText:
                  type: string
      responses:
        '200':
          description: Image updated successfully

  # Tag Routes
...